Bug fix for the amount issue

+++ razorpay-quick-payments.php
@@ -4,7 +4,7 @@
  * Plugin Name: Razorpay Quick Payments
  * Plugin URI: https://github.com/razorpay/razorpay-quick-payments
  * Description: Quick Payments for Wordpress, by Razorpay.
- * Version: 1.2.0
+ * Version: 1.2.1
  * Author: Team Razorpay
  * Author URI: https://razorpay.com/about/
  * License: GPL2
@@ -104,7 +104,7 @@ function wordpressRazorpayInit()
 
             $metadata = get_post_meta($pageID);
 
-            $amount = (int) ($metadata['amount'][0]) * 100;
+            $amount = ($metadata['amount'][0]) * 100;
 
             if (isset($this->keyID) && isset($this->keySecret) && $amount!=null)
             {
@@ -143,7 +143,7 @@ function wordpressRazorpayInit()
                 }
                 else
                 {
-                    $amount = (int) ($metadata['amount'][0]) * 100;
+                    $amount = ($metadata['amount'][0]) * 100;
 
                     $productInfo = $this->getProductDecription($metadata, $pageID);
